+++ CHANGELOG.md
@@ -4077,3 +4077,4 @@
 - [2026-08-31 14:28] test(test): implement widget integration tests for test
 - [2026-08-31 15:50] test(network): add unit tests for network business logic
 - [2026-08-31 16:04] feat(core): add bi-directional event bus handler for core
+- [2026-08-31 16:15] fix(security): prevent double tap duplicate action in security